requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
bottle==0.12.19
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
crypto==1.4.1
Deprecated==1.2.13
Django==4.0.3
django-appconf==1.0.5
django-compressor==3.1
django-haystack==3.2.dev0
django-ipware==4.0.2
django-mdeditor==0.1.20
django-uuslug==2.0.0
djangorestframework==3.14.0
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
hashids==1.3.1
idna==3.3
importlib-metadata==4.11.3
jieba==0.42.1
jsonpickle==2.1.0
Markdown==3.3.6
mysqlclient==2.1.0
Naked==0.1.31
packaging==21.3
Pillow==9.0.1
pycryptodome==3.14.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-logstash==0.4.6
python-memcached==1.59
python-slugify==6.1.1
pytz==2022.1
PyYAML==6.0
raven==6.10.0
rcssmin==1.1.0
redis==4.1.4
requests==2.27.1
rjsmin==1.2.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
webencodings==0.5.1
WeRoBot==1.13.1
Whoosh==2.7.4
wrapt==1.14.0
xmltodict==0.12.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0